+++ lib/widgets/reorderable_animated_positioned.dart
@@ -30,14 +30,20 @@ class ReorderableAnimatedPositioned extends StatefulWidget {
   final VoidCallback onMovingFinished;
 
   /// Duration for the position change of [child] (won't be used while dragging!).
-  final Duration positionDuration;
+  final Duration positionChangeDuration;
+
+  /// Default duration when an item changes his position.
+  ///
+  /// This duration will be used for position changes while dragging.
+  final Duration draggingPositionChangeDuration;
 
   const ReorderableAnimatedPositioned({
     required this.child,
     required this.reorderableEntity,
     required this.isDragging,
     required this.onMovingFinished,
-    required this.positionDuration,
+    required this.positionChangeDuration,
+    required this.draggingPositionChangeDuration,
     Key? key,
   }) : super(key: key);
 
@@ -49,12 +55,6 @@ class ReorderableAnimatedPositioned extends StatefulWidget {
 class _ReorderableAnimatedPositionedState
     extends State<ReorderableAnimatedPositioned>
     with SingleTickerProviderStateMixin {
-  /// Default duration when an item changes his position.
-  ///
-  /// This duration will be used for position changes while draggong or not
-  /// dragging.
-  final _defaultAnimationDuration = const Duration(milliseconds: 200);
-
   /// This controller will be used for the animation when the position changes.
   late AnimationController _animationController;
 
@@ -66,7 +66,7 @@ class _ReorderableAnimatedPositionedState
     super.initState();
 
     _animationController = AnimationController(
-      duration: _defaultAnimationDuration,
+      duration: widget.draggingPositionChangeDuration,
       vsync: this,
     );
 
@@ -155,7 +155,7 @@ class _ReorderableAnimatedPositionedState
     Offset begin = Offset.zero,
     Offset end = Offset.zero,
   }) async {
-    _animationController.duration = _defaultAnimationDuration;
+    _animationController.duration = widget.draggingPositionChangeDuration;
     final tween = Tween<Offset>(begin: begin, end: end);
     _offsetAnimation = tween.animate(_animationController)
       ..addListener(() {
@@ -184,7 +184,7 @@ class _ReorderableAnimatedPositionedState
   /// Important, this function is only added when [widget.child] updates his
   /// position, not while dragging.
   Future<void> _animateOffset({required Offset begin}) async {
-    _animationController.duration = widget.positionDuration;
+    _animationController.duration = widget.positionChangeDuration;
 
     final tween = Tween<Offset>(begin: begin, end: Offset.zero);
     _offsetAnimation = tween.animate(_animationController)